+++ ssolv-api-place/src/test/kotlin/org/depromeet/team3/place/application/plan/SelectSurveyKeywordsServiceTest.kt
@@ -1,8 +1,8 @@
 package org.depromeet.team3.place.application.plan
 
 import org.assertj.core.api.Assertions.assertThat
-import org.depromeet.team3.meeting.MeetingQuery
-import org.depromeet.team3.place.application.model.PlaceSurveySummary
+import org.depromeet.team3.place.application.search.model.PlaceSurveySummary
+import org.depromeet.team3.place.application.search.planner.SelectSurveyKeywordsService
 import org.depromeet.team3.surveycategory.SurveyCategory
 import org.depromeet.team3.surveycategory.SurveyCategoryLevel
 import org.junit.jupiter.api.BeforeEach
